Hot fixes (including scripts)
