+++ Task_1_1_1/src/junit_console/junit.sh
@@ -1,2 +1,2 @@
-javac -cp "junit-code-platform-standalone-1.11.0.jar" ../main/java/ru/nsu/dubrovin/Main.java ../test/java/ru/nsu/dubrovin/MainTest.java -d "."
-java -jar junit-platform-console-standalone-1.11.0.jar execute -p ru.nsu.dubrovin -cp ru/nsu/dubrovin -cp . --select-class ru.nsu.dubrovin.MainTest --reports-dir='reports'
+javac -cp "junit-platform-console-standalone-1.11.0.jar" ../main/java/ru/nsu/dubrovin/Heapsort.java ../test/java/ru/nsu/dubrovin/HeapsortTest.java -d "."
+java -jar junit-platform-console-standalone-1.11.0.jar execute -p ru.nsu.dubrovin -cp ru/nsu/dubrovin -cp . --select-class ru.nsu.dubrovin.HeapsortTest --reports-dir='reports'
