+++ Task_1_1_1/src/junit_console/reports/TEST-junit-jupiter.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<testsuite name="JUnit Jupiter" tests="10" skipped="0" failures="0" errors="0" time="0.051" hostname="DESKTOP-VANO1TF" timestamp="2024-09-08T22:37:48">
+<testsuite name="JUnit Jupiter" tests="10" skipped="0" failures="0" errors="0" time="0.052" hostname="DESKTOP-VANO1TF" timestamp="2024-09-08T23:59:37">
 <properties>
 <property name="file.encoding" value="UTF-8"/>
 <property name="file.separator" value="\"/>
@@ -42,7 +42,7 @@
 <property name="sun.cpu.endian" value="little"/>
 <property name="sun.cpu.isalist" value="amd64"/>
 <property name="sun.io.unicode.encoding" value="UnicodeLittle"/>
-<property name="sun.java.command" value="junit-platform-console-standalone-1.11.0.jar execute -p ru.nsu.dubrovin -cp ru/nsu/dubrovin -cp . --select-class ru.nsu.dubrovin.MainTest --reports-dir=reports"/>
+<property name="sun.java.command" value="junit-platform-console-standalone-1.11.0.jar execute -p ru.nsu.dubrovin -cp ru/nsu/dubrovin -cp . --select-class ru.nsu.dubrovin.HeapsortTest --reports-dir=reports"/>
 <property name="sun.java.launcher" value="SUN_STANDARD"/>
 <property name="sun.jnu.encoding" value="Cp1251"/>
 <property name="sun.management.compiler" value="HotSpot 64-Bit Tiered Compilers"/>
@@ -56,63 +56,63 @@
 <property name="user.timezone" value="Asia/Novosibirsk"/>
 <property name="user.variant" value=""/>
 </properties>
-<testcase name="test0EmptyArray()" classname="ru.nsu.dubrovin.MainTest" time="0.001">
+<testcase name="test3SortedArray()" classname="ru.nsu.dubrovin.HeapsortTest" time="0.002">
 <system-out><![CDATA[
-unique-id: [engine:junit-jupiter]/[class:ru.nsu.dubrovin.MainTest]/[method:test0EmptyArray()]
-display-name: test0EmptyArray()
-]]></system-out>
-</testcase>
-<testcase name="test5AlreadyHeap()" classname="ru.nsu.dubrovin.MainTest" time="0">
-<system-out><![CDATA[
-unique-id: [engine:junit-jupiter]/[class:ru.nsu.dubrovin.MainTest]/[method:test5AlreadyHeap()]
-display-name: test5AlreadyHeap()
-]]></system-out>
-</testcase>
-<testcase name="teat1OneElement()" classname="ru.nsu.dubrovin.MainTest" time="0">
-<system-out><![CDATA[
-unique-id: [engine:junit-jupiter]/[class:ru.nsu.dubrovin.MainTest]/[method:teat1OneElement()]
-display-name: teat1OneElement()
-]]></system-out>
-</testcase>
-<testcase name="test6Random()" classname="ru.nsu.dubrovin.MainTest" time="0">
-<system-out><![CDATA[
-unique-id: [engine:junit-jupiter]/[class:ru.nsu.dubrovin.MainTest]/[method:test6Random()]
-display-name: test6Random()
-]]></system-out>
-</testcase>
-<testcase name="test8IntBorders()" classname="ru.nsu.dubrovin.MainTest" time="0">
-<system-out><![CDATA[
-unique-id: [engine:junit-jupiter]/[class:ru.nsu.dubrovin.MainTest]/[method:test8IntBorders()]
-display-name: test8IntBorders()
-]]></system-out>
-</testcase>
-<testcase name="timeTest()" classname="ru.nsu.dubrovin.MainTest" time="0.017">
-<system-out><![CDATA[
-unique-id: [engine:junit-jupiter]/[class:ru.nsu.dubrovin.MainTest]/[method:timeTest()]
-display-name: timeTest()
-]]></system-out>
-</testcase>
-<testcase name="test3SortedArray()" classname="ru.nsu.dubrovin.MainTest" time="0.001">
-<system-out><![CDATA[
-unique-id: [engine:junit-jupiter]/[class:ru.nsu.dubrovin.MainTest]/[method:test3SortedArray()]
+unique-id: [engine:junit-jupiter]/[class:ru.nsu.dubrovin.HeapsortTest]/[method:test3SortedArray()]
 display-name: test3SortedArray()
 ]]></system-out>
 </testcase>
-<testcase name="test2SameElements()" classname="ru.nsu.dubrovin.MainTest" time="0.001">
+<testcase name="timeTest()" classname="ru.nsu.dubrovin.HeapsortTest" time="0.016">
 <system-out><![CDATA[
-unique-id: [engine:junit-jupiter]/[class:ru.nsu.dubrovin.MainTest]/[method:test2SameElements()]
+unique-id: [engine:junit-jupiter]/[class:ru.nsu.dubrovin.HeapsortTest]/[method:timeTest()]
+display-name: timeTest()
+]]></system-out>
+</testcase>
+<testcase name="test0EmptyArray()" classname="ru.nsu.dubrovin.HeapsortTest" time="0">
+<system-out><![CDATA[
+unique-id: [engine:junit-jupiter]/[class:ru.nsu.dubrovin.HeapsortTest]/[method:test0EmptyArray()]
+display-name: test0EmptyArray()
+]]></system-out>
+</testcase>
+<testcase name="test5AlreadyHeap()" classname="ru.nsu.dubrovin.HeapsortTest" time="0.001">
+<system-out><![CDATA[
+unique-id: [engine:junit-jupiter]/[class:ru.nsu.dubrovin.HeapsortTest]/[method:test5AlreadyHeap()]
+display-name: test5AlreadyHeap()
+]]></system-out>
+</testcase>
+<testcase name="test8IntBorders()" classname="ru.nsu.dubrovin.HeapsortTest" time="0">
+<system-out><![CDATA[
+unique-id: [engine:junit-jupiter]/[class:ru.nsu.dubrovin.HeapsortTest]/[method:test8IntBorders()]
+display-name: test8IntBorders()
+]]></system-out>
+</testcase>
+<testcase name="test2SameElements()" classname="ru.nsu.dubrovin.HeapsortTest" time="0">
+<system-out><![CDATA[
+unique-id: [engine:junit-jupiter]/[class:ru.nsu.dubrovin.HeapsortTest]/[method:test2SameElements()]
 display-name: test2SameElements()
 ]]></system-out>
 </testcase>
-<testcase name="test7Random()" classname="ru.nsu.dubrovin.MainTest" time="0">
+<testcase name="test6Random()" classname="ru.nsu.dubrovin.HeapsortTest" time="0">
 <system-out><![CDATA[
-unique-id: [engine:junit-jupiter]/[class:ru.nsu.dubrovin.MainTest]/[method:test7Random()]
+unique-id: [engine:junit-jupiter]/[class:ru.nsu.dubrovin.HeapsortTest]/[method:test6Random()]
+display-name: test6Random()
+]]></system-out>
+</testcase>
+<testcase name="teat1OneElement()" classname="ru.nsu.dubrovin.HeapsortTest" time="0">
+<system-out><![CDATA[
+unique-id: [engine:junit-jupiter]/[class:ru.nsu.dubrovin.HeapsortTest]/[method:teat1OneElement()]
+display-name: teat1OneElement()
+]]></system-out>
+</testcase>
+<testcase name="test7Random()" classname="ru.nsu.dubrovin.HeapsortTest" time="0">
+<system-out><![CDATA[
+unique-id: [engine:junit-jupiter]/[class:ru.nsu.dubrovin.HeapsortTest]/[method:test7Random()]
 display-name: test7Random()
 ]]></system-out>
 </testcase>
-<testcase name="test4ReversedSortedArray()" classname="ru.nsu.dubrovin.MainTest" time="0">
+<testcase name="test4ReversedSortedArray()" classname="ru.nsu.dubrovin.HeapsortTest" time="0">
 <system-out><![CDATA[
-unique-id: [engine:junit-jupiter]/[class:ru.nsu.dubrovin.MainTest]/[method:test4ReversedSortedArray()]
+unique-id: [engine:junit-jupiter]/[class:ru.nsu.dubrovin.HeapsortTest]/[method:test4ReversedSortedArray()]
 display-name: test4ReversedSortedArray()
 ]]></system-out>
 </testcase>
