+++ Task_1_1_1/src/junit_console/reports/TEST-junit-vintage.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<testsuite name="JUnit Vintage" tests="0" skipped="0" failures="0" errors="0" time="0.001" hostname="DESKTOP-VANO1TF" timestamp="2024-09-08T22:37:49">
+<testsuite name="JUnit Vintage" tests="0" skipped="0" failures="0" errors="0" time="0.001" hostname="DESKTOP-VANO1TF" timestamp="2024-09-08T23:59:37">
 <properties>
 <property name="file.encoding" value="UTF-8"/>
 <property name="file.separator" value="\"/>
@@ -42,7 +42,7 @@
 <property name="sun.cpu.endian" value="little"/>
 <property name="sun.cpu.isalist" value="amd64"/>
 <property name="sun.io.unicode.encoding" value="UnicodeLittle"/>
-<property name="sun.java.command" value="junit-platform-console-standalone-1.11.0.jar execute -p ru.nsu.dubrovin -cp ru/nsu/dubrovin -cp . --select-class ru.nsu.dubrovin.MainTest --reports-dir=reports"/>
+<property name="sun.java.command" value="junit-platform-console-standalone-1.11.0.jar execute -p ru.nsu.dubrovin -cp ru/nsu/dubrovin -cp . --select-class ru.nsu.dubrovin.HeapsortTest --reports-dir=reports"/>
 <property name="sun.java.launcher" value="SUN_STANDARD"/>
 <property name="sun.jnu.encoding" value="Cp1251"/>
 <property name="sun.management.compiler" value="HotSpot 64-Bit Tiered Compilers"/>
