--- Task_1_1_1/src/main/java/element-list
@@ -1,1 +1,0 @@
-ru.nsu.dubrovin
