--- Task_1_1_1/src/main/java/index-all.html
@@ -1,82 +1,0 @@
-<!DOCTYPE HTML>
-<html lang="ru">
-<head>
-<!-- Generated by javadoc (21) on Sun Sep 08 21:57:01 NOVT 2024 -->
-<title>Index</title>
-<meta name="viewport" content="width=device-width, initial-scale=1">
-<meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-<meta name="dc.created" content="2024-09-08">
-<meta name="description" content="index">
-<meta name="generator" content="javadoc/IndexWriter">
-<link rel="stylesheet" type="text/css" href="stylesheet.css" title="Style">
-<link rel="stylesheet" type="text/css" href="script-dir/jquery-ui.min.css" title="Style">
-<script type="text/javascript" src="script.js"></script>
-<script type="text/javascript" src="script-dir/jquery-3.6.1.min.js"></script>
-<script type="text/javascript" src="script-dir/jquery-ui.min.js"></script>
-</head>
-<body class="index-page">
-<script type="text/javascript">var pathtoroot = "./";
-loadScripts(document, 'script');</script>
-<noscript>
-<div>JavaScript is disabled on your browser.</div>
-</noscript>
-<div class="flex-box">
-<header role="banner" class="flex-header">
-<nav role="navigation">
-<!-- ========= START OF TOP NAVBAR ======= -->
-<div class="top-nav" id="navbar-top"><button id="navbar-toggle-button" aria-controls="navbar-top" aria-expanded="false" aria-label="Toggle navigation links"><span class="nav-bar-toggle-icon">&nbsp;</span><span class="nav-bar-toggle-icon">&nbsp;</span><span class="nav-bar-toggle-icon">&nbsp;</span></button>
-<div class="skip-nav"><a href="#skip-navbar-top" title="Skip navigation links">Skip navigation links</a></div>
-<ul id="navbar-top-firstrow" class="nav-list" title="Navigation">
-<li><a href="ru/nsu/dubrovin/package-summary.html">Package</a></li>
-<li>Class</li>
-<li><a href="ru/nsu/dubrovin/package-tree.html">Tree</a></li>
-<li class="nav-bar-cell1-rev">Index</li>
-<li><a href="help-doc.html#index">Help</a></li>
-</ul>
-</div>
-<div class="sub-nav">
-<div id="navbar-sub-list"></div>
-<div class="nav-list-search"><a href="search.html">SEARCH</a>
-<input type="text" id="search-input" disabled placeholder="Search">
-<input type="reset" id="reset-button" disabled value="reset">
-</div>
-</div>
-<!-- ========= END OF TOP NAVBAR ========= -->
-<span class="skip-nav" id="skip-navbar-top"></span></nav>
-</header>
-<div class="flex-content">
-<main role="main">
-<div class="header">
-<h1>Index</h1>
-</div>
-<a href="#I:H">H</a>&nbsp;<a href="#I:M">M</a>&nbsp;<a href="#I:R">R</a>&nbsp;<br><a href="allclasses-index.html">All&nbsp;Classes&nbsp;and&nbsp;Interfaces</a><span class="vertical-separator">|</span><a href="allpackages-index.html">All&nbsp;Packages</a>
-<h2 class="title" id="I:H">H</h2>
-<dl class="index">
-<dt><a href="ru/nsu/dubrovin/Main.html#heapsort(int%5B%5D)" class="member-name-link">heapsort(int[])</a> - Static method in class ru.nsu.dubrovin.<a href="ru/nsu/dubrovin/Main.html" title="class in ru.nsu.dubrovin">Main</a></dt>
-<dd>
-<div class="block">Сортировка двоичной кучей.</div>
-</dd>
-</dl>
-<h2 class="title" id="I:M">M</h2>
-<dl class="index">
-<dt><a href="ru/nsu/dubrovin/Main.html#main(java.lang.String%5B%5D)" class="member-name-link">main(String[])</a> - Static method in class ru.nsu.dubrovin.<a href="ru/nsu/dubrovin/Main.html" title="class in ru.nsu.dubrovin">Main</a></dt>
-<dd>
-<div class="block">Главный метод (не используется).</div>
-</dd>
-<dt><a href="ru/nsu/dubrovin/Main.html" class="type-name-link" title="class in ru.nsu.dubrovin">Main</a> - Class in <a href="ru/nsu/dubrovin/package-summary.html">ru.nsu.dubrovin</a></dt>
-<dd>
-<div class="block">Основной класс.</div>
-</dd>
-<dt><a href="ru/nsu/dubrovin/Main.html#%3Cinit%3E()" class="member-name-link">Main()</a> - Constructor for class ru.nsu.dubrovin.<a href="ru/nsu/dubrovin/Main.html" title="class in ru.nsu.dubrovin">Main</a></dt>
-<dd>&nbsp;</dd>
-</dl>
-<h2 class="title" id="I:R">R</h2>
-<dl class="index">
-<dt><a href="ru/nsu/dubrovin/package-summary.html">ru.nsu.dubrovin</a> - package ru.nsu.dubrovin</dt>
-<dd>&nbsp;</dd>
-</dl>
-<a href="#I:H">H</a>&nbsp;<a href="#I:M">M</a>&nbsp;<a href="#I:R">R</a>&nbsp;<br><a href="allclasses-index.html">All&nbsp;Classes&nbsp;and&nbsp;Interfaces</a><span class="vertical-separator">|</span><a href="allpackages-index.html">All&nbsp;Packages</a></main>
-</div>
-</div>
-</body>
-</html>
