--- Task_1_1_1/src/main/java/index.html
@@ -1,26 +1,0 @@
-<!DOCTYPE HTML>
-<html lang="ru">
-<head>
-<!-- Generated by javadoc (21) on Sun Sep 08 21:57:01 NOVT 2024 -->
-<title>Generated Documentation (Untitled)</title>
-<meta name="viewport" content="width=device-width, initial-scale=1">
-<meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-<meta name="dc.created" content="2024-09-08">
-<meta name="description" content="index redirect">
-<meta name="generator" content="javadoc/IndexRedirectWriter">
-<link rel="canonical" href="ru/nsu/dubrovin/package-summary.html">
-<link rel="stylesheet" type="text/css" href="stylesheet.css" title="Style">
-<script type="text/javascript">window.location.replace('ru/nsu/dubrovin/package-summary.html')</script>
-<noscript>
-<meta http-equiv="Refresh" content="0;ru/nsu/dubrovin/package-summary.html">
-</noscript>
-</head>
-<body class="index-redirect-page">
-<main role="main">
-<noscript>
-<p>JavaScript is disabled on your browser.</p>
-</noscript>
-<p><a href="ru/nsu/dubrovin/package-summary.html">ru/nsu/dubrovin/package-summary.html</a></p>
-</main>
-</body>
-</html>
