--- Task_1_1_1/src/main/java/legal/COPYRIGHT
@@ -1,1 +1,0 @@
-Please see ..\java.base\COPYRIGHT
