--- Task_1_1_1/src/main/java/legal/LICENSE
@@ -1,1 +1,0 @@
-Please see ..\java.base\LICENSE
