--- Task_1_1_1/src/main/java/member-search-index.js
@@ -1,1 +1,0 @@
-memberSearchIndex = [{"p":"ru.nsu.dubrovin","c":"Main","l":"heapsort(int[])"},{"p":"ru.nsu.dubrovin","c":"Main","l":"Main()","u":"%3Cinit%3E()"},{"p":"ru.nsu.dubrovin","c":"Main","l":"main(String[])","u":"main(java.lang.String[])"}];updateSearchResults();
