--- Task_1_1_1/src/main/java/module-search-index.js
@@ -1,1 +1,0 @@
-moduleSearchIndex = [];updateSearchResults();
