--- Task_1_1_1/src/main/java/package-search-index.js
@@ -1,1 +1,0 @@
-packageSearchIndex = [{"l":"All Packages","u":"allpackages-index.html"},{"l":"ru.nsu.dubrovin"}];updateSearchResults();
