+++ Task_1_1_1/src/main/java/ru/nsu/dubrovin/Heapsort.java
@@ -1,9 +1,9 @@
 package ru.nsu.dubrovin;
 
 /**
- * Основной класс.
+ * Основной класс, содержит публичный метод сортировки кучей.
  */
-public class Main {
+public class Heapsort {
 
     /**
      * Главный метод (не используется).
@@ -26,7 +26,7 @@ public class Main {
      *
      * @return глубина рекурсии
      */
-    private static int heapify(int[] arr, int n, int i, int[] depth) {
+    private static int heapify(int[] arr, int n, int i, int depth) {
         int largest = i;
         int left = i * 2 + 1;
         int right = i * 2 + 2;
@@ -41,11 +41,10 @@ public class Main {
 
         if (largest != i) {
             swap(arr, i, largest);
-            depth[0] += 1;
-            heapify(arr, n, largest, depth);
+            depth = heapify(arr, n, largest, depth);
         }
 
-        return depth[0];
+        return depth + 1;
     }
 
     /**
@@ -53,14 +52,14 @@ public class Main {
      *
      * @param arr сам массив
      *
-     * @param fstind индекс первого элемента
+     * @param fstInd индекс первого элемента
      *
-     * @param sndind индекс второго элемента
+     * @param sndInd индекс второго элемента
      */
-    private static void swap(int[] arr, int fstind, int sndind) {
-        int tmp = arr[fstind];
-        arr[fstind] = arr[sndind];
-        arr[sndind] = tmp;
+    private static void swap(int[] arr, int fstInd, int sndInd) {
+        int tmp = arr[fstInd];
+        arr[fstInd] = arr[sndInd];
+        arr[sndInd] = tmp;
     }
 
     /**
@@ -74,12 +73,12 @@ public class Main {
         int len = arr.length;
         int iterations = 0;
         for (int i = len / 2 - 1; i >= 0; i--) {
-            iterations += heapify(arr, len, i, new int[]{1});
+            iterations += heapify(arr, len, i, 0);
         }
 
         for (int i = len - 1; i >= 0; i--) {
             swap(arr, 0, i);
-            iterations += heapify(arr, i, 0, new int[]{1});
+            iterations += heapify(arr, i, 0, 0);
         }
 
         return iterations;
