--- Task_1_1_1/src/main/java/ru/nsu/dubrovin/package-summary.html
@@ -1,92 +1,0 @@
-<!DOCTYPE HTML>
-<html lang="ru">
-<head>
-<!-- Generated by javadoc (21) on Sun Sep 08 21:57:01 NOVT 2024 -->
-<title>ru.nsu.dubrovin</title>
-<meta name="viewport" content="width=device-width, initial-scale=1">
-<meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-<meta name="dc.created" content="2024-09-08">
-<meta name="description" content="declaration: package: ru.nsu.dubrovin">
-<meta name="generator" content="javadoc/PackageWriterImpl">
-<link rel="stylesheet" type="text/css" href="../../../stylesheet.css" title="Style">
-<link rel="stylesheet" type="text/css" href="../../../script-dir/jquery-ui.min.css" title="Style">
-<script type="text/javascript" src="../../../script.js"></script>
-<script type="text/javascript" src="../../../script-dir/jquery-3.6.1.min.js"></script>
-<script type="text/javascript" src="../../../script-dir/jquery-ui.min.js"></script>
-</head>
-<body class="package-declaration-page">
-<script type="text/javascript">var pathtoroot = "../../../";
-loadScripts(document, 'script');</script>
-<noscript>
-<div>JavaScript is disabled on your browser.</div>
-</noscript>
-<div class="flex-box">
-<header role="banner" class="flex-header">
-<nav role="navigation">
-<!-- ========= START OF TOP NAVBAR ======= -->
-<div class="top-nav" id="navbar-top"><button id="navbar-toggle-button" aria-controls="navbar-top" aria-expanded="false" aria-label="Toggle navigation links"><span class="nav-bar-toggle-icon">&nbsp;</span><span class="nav-bar-toggle-icon">&nbsp;</span><span class="nav-bar-toggle-icon">&nbsp;</span></button>
-<div class="skip-nav"><a href="#skip-navbar-top" title="Skip navigation links">Skip navigation links</a></div>
-<ul id="navbar-top-firstrow" class="nav-list" title="Navigation">
-<li class="nav-bar-cell1-rev">Package</li>
-<li>Class</li>
-<li><a href="package-tree.html">Tree</a></li>
-<li><a href="../../../index-all.html">Index</a></li>
-<li><a href="../../../help-doc.html#package">Help</a></li>
-</ul>
-<ul class="sub-nav-list-small">
-<li>
-<p>Package:</p>
-<ul>
-<li>Description</li>
-<li>Related Packages</li>
-<li><a href="#class-summary">Classes and Interfaces</a></li>
-</ul>
-</li>
-</ul>
-</div>
-<div class="sub-nav">
-<div id="navbar-sub-list">
-<ul class="sub-nav-list">
-<li>Package:&nbsp;</li>
-<li>Description&nbsp;|&nbsp;</li>
-<li>Related Packages&nbsp;|&nbsp;</li>
-<li><a href="#class-summary">Classes and Interfaces</a></li>
-</ul>
-</div>
-<div class="nav-list-search"><a href="../../../search.html">SEARCH</a>
-<input type="text" id="search-input" disabled placeholder="Search">
-<input type="reset" id="reset-button" disabled value="reset">
-</div>
-</div>
-<!-- ========= END OF TOP NAVBAR ========= -->
-<span class="skip-nav" id="skip-navbar-top"></span></nav>
-</header>
-<div class="flex-content">
-<main role="main">
-<div class="header">
-<h1 title="Package ru.nsu.dubrovin" class="title">Package ru.nsu.dubrovin</h1>
-</div>
-<hr>
-<div class="package-signature">package <span class="element-name">ru.nsu.dubrovin</span></div>
-<section class="summary">
-<ul class="summary-list">
-<li>
-<div id="class-summary">
-<div class="caption"><span>Classes</span></div>
-<div class="summary-table two-column-summary">
-<div class="table-header col-first">Class</div>
-<div class="table-header col-last">Description</div>
-<div class="col-first even-row-color class-summary class-summary-tab2"><a href="Main.html" title="class in ru.nsu.dubrovin">Main</a></div>
-<div class="col-last even-row-color class-summary class-summary-tab2">
-<div class="block">Основной класс.</div>
-</div>
-</div>
-</div>
-</li>
-</ul>
-</section>
-</main>
-</div>
-</div>
-</body>
-</html>
