+++ Task_1_1_1/src/main/java/script.sh
@@ -1,4 +1,4 @@
-javac ru/nsu/dubrovin/Main.java #Компилим исходный код в байткод JVM
-jar cfe heapsort.jar ru.nsu.dubrovin.Main ru/nsu/dubrovin/Main.class #Собираем jar-file
+javac ru/nsu/dubrovin/Heapsort.java #Компилим исходный код в байткод JVM
+jar cfe heapsort.jar ru.nsu.dubrovin.Heapsort ru/nsu/dubrovin/Heapsort.class #Собираем jar-file
 java -jar heapsort.jar #Запускаем jar-file
 javadoc ru.nsu.dubrovin #Генерим джавадок
