--- Task_1_1_1/src/main/java/tag-search-index.js
@@ -1,1 +1,0 @@
-tagSearchIndex = [];updateSearchResults();
