--- Task_1_1_1/src/main/java/type-search-index.js
@@ -1,1 +1,0 @@
-typeSearchIndex = [{"l":"All Classes and Interfaces","u":"allclasses-index.html"},{"p":"ru.nsu.dubrovin","l":"Main"}];updateSearchResults();
